+++ docs/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.2.2] – 2026-08-23
+
+### Changed
+- Version bump to `1.2.2`.
+
+---
 ## [1.2.1] – 2026-08-23
 
 ### Changed